Deleted most of the old, single connection code
