+++ SinusSynchronous/Services/Mediator/Messages.cs
@@ -56,12 +56,6 @@ public record CharacterDataCreatedMessage(CharacterData CharacterData) : SameThr
 public record CharacterDataAnalyzedMessage : MessageBase;
 public record PenumbraStartRedrawMessage(IntPtr Address) : MessageBase;
 public record PenumbraEndRedrawMessage(IntPtr Address) : MessageBase;
-[Obsolete("no longer needed for multiconnect")]
-public record HubReconnectingMessage(Exception? Exception) : SameThreadMessage;
-[Obsolete("no longer needed for multiconnect")]
-public record HubReconnectedMessage(string? Arg) : SameThreadMessage;
-[Obsolete("no longer needed for multiconnect")]
-public record HubClosedMessage(Exception? Exception) : SameThreadMessage;
 public record DownloadReadyMessage(Guid RequestId) : MessageBase;
 public record DownloadStartedMessage(GameObjectHandler DownloadId, Dictionary<string, FileDownloadStatus> DownloadStatus) : MessageBase;
 public record DownloadFinishedMessage(GameObjectHandler DownloadId) : MessageBase;
