+++ SinusSynchronous/SinusSynchronous.csproj
@@ -80,7 +80,6 @@
 
   <ItemGroup>
     <Folder Include="MultiConnectAPI\" />
-    <Folder Include="WebAPI\SignalR\Client\" />
   </ItemGroup>
 
 </Project>
